app/requirements.txt
uvicorn==0.22.0
fastapi==0.98.0
pydantic==1.10.9
torch==2.0.1
torchvision==0.15.2
faiss-cpu==1.7.4
pandas==2.0.2
tqdm==4.65.0
scikit-learn==1.3.0
python-multipart==0.0.6